refactor: reorganize folder structure for readability (issue #111)
Consolidates 7 top-level folders + platform/web/ subdirectory into 5 clear layers:

- controllers/ + platform/web/ → hooks/ (React hooks with app logic) and readers/ (all file format readers)
- platform/web/htmlParser.js renamed to readers/readHtml.js (consistent read* naming)
- presentation/ → ui/presenters/ (co-located with its only consumers)
- ui/ResultTabs/ flattened into ui/ (removes unnecessary subfolder depth)

Only file moves and import updates — no logic changes.
