+++ src/ui/DividendsTab.jsx
@@ -1,9 +1,9 @@
-import { t } from '../../localization/i18n.js'
+import { t } from '../localization/i18n.js'
 import { Box, Typography } from '@mui/material'
 import { alpha } from '@mui/material/styles'
 import { ReceiptLongOutlined } from '@mui/icons-material'
 import DataTable from './DataTable.jsx'
-import { DividendPresenter } from '../../presentation/DividendPresenter.js'
+import { DividendPresenter } from './presenters/DividendPresenter.js'
 
 export default function DividendsTab({ result }) {
   // --- DIVIDENDS (presentation) ---
