+++ src/ui/ExcelCopyButton.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
-import { t } from '../../localization/i18n.js'
+import { t } from '../localization/i18n.js'
 import { Button, Tooltip } from '@mui/material'
 import { ContentCopyOutlined, CheckOutlined, ErrorOutlineOutlined } from '@mui/icons-material'
-import { ExcelPresenter } from '../../presentation/ExcelPresenter.js'
+import { ExcelPresenter } from './presenters/ExcelPresenter.js'
 
 const RESET_DELAY = 2500
