+++ src/ui/TaxableToggleDialog.jsx
@@ -1,6 +1,6 @@
 import { Dialog, DialogTitle, DialogContent, DialogActions, Button, Typography, IconButton } from '@mui/material'
 import { Close } from '@mui/icons-material'
-import { t } from '../../localization/i18n.js'
+import { t } from '../localization/i18n.js'
 
 export default function TaxableToggleDialog({ pending, onClose, onConfirm }) {
