+++ src/ui/presenters/DividendPresenter.js
@@ -1,7 +1,7 @@
 // DividendPresenter.js
-import { t } from '../localization/i18n.js'
+import { t } from '../../localization/i18n.js'
 import { fmt } from './fmt.js'
-import { decimalToNumber } from '../core/domain/numStr.js'
+import { decimalToNumber } from '../../core/domain/numStr.js'
 
 export class DividendPresenter {
   constructor({ lcl, mode = 'display' }) {
